Added gestureDrawing function to new file
